Author: Hannes Borch Date: 04-07-09
SBMLsqueezer now checks for updates. If a newer version of SBMLsqueezer is available online, the user is notified by a message window in the bottom right corner of the screen. Furthermore, the release notes of the most recent SBMLsqueezer version are shown to the user by clicking on "show release notes".

For that, the initialization of the SBMLsqueezerUI objects was relocated in a class extended from java.lang.Thread in order to not influence SBMLsqueeezer's performance by checking for updates.